Usando classes
código reformulado e mais organizado usando classes
